pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "material-db"
version = "0.1.0"
description = "Open Material Database for 3D printing materials"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "PyYAML>=6.0",
    "jsonschema>=4.0.0",
    "referencing>=0.37.0",
    "requests>=2.31.0",
    "google-cloud-storage>=2.10.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["scripts*"]

[tool.setuptools.package-data]
scripts = ["*.py"]
